added prod to the template fro db.

+++ stdPatterns/newTSRepo.sh
@@ -305,10 +305,13 @@ EOL
     cat >> src/database.ts << EOL
 import { createConnection } from 'typeorm';
 import { User } from './entity/user';
+import { __prod__ } from './constants'
 
 export const init_db = async() => {
   const connection = await createConnection();
-  await connection.dropDatabase();
+  if (!__prod__) {
+    await connection.dropDatabase();
+  }
   await connection.synchronize();
 
   // User
